added lesson model and controller

+++ backend/controllers/moduleController.js
@@ -1,6 +1,8 @@
 import Module from "../models/Module.js";
+import Lesson from "../models/Lesson.js";
 import cloudinary from "../config/cloudinary.js";
 import streamifier from "streamifier";
+
 
 // TODO: createModule, getModules, getModuleById, updateModule, publishModule, deleteModule
 
@@ -18,24 +20,23 @@ const uploadToCloudinary = (fileBuffer) => {
     }); 
 };
 
+const deleteFromCloudinary = async (publicId) => {
+    if (!publicId) return;
+
+    await cloudinary.uploader.destroy(publicId, {
+        resource_type: "video"
+    })
+}
+
 // New Module
 export const createModule = async (req, res) =>{
     try {
         const {title, description, reflectionQuestion, duration} = req.body
-        let videoUrl;
-        let videoPublicId;
-
-        if (req.file) {
-            const uploadResult = await uploadToCloudinary(req.file.buffer);
-            videoUrl = uploadResult.secure_url
-            videoPublicId = uploadResult.public_id
-        }
-
+                
         const newModule = await Module.create({
             title,
             description,
-            videoUrl,
-            videoPublicId,
+            createdBy: req.user.id,
             reflectionQuestion,
             duration
         })
@@ -47,6 +48,62 @@ export const createModule = async (req, res) =>{
         return res
         .status(500)
         .json({message: "error creating module", error: error.message})
+    }
+}
+
+export const addLessonToModule = async (req, res) => {
+    try {
+        const {moduleId} = req.params;
+        const {title, overview, duration, order} = req.body;
+
+        
+        const module = await Module.findById(moduleId);
+        if (!module) {
+            return res
+            .status(404).json({message: "module not found"})
+        };
+
+        if(!req.file) {
+            return res.status(400).json({message: "Lesson video is required"})
+        }
+
+        const uploadResult = await uploadToCloudinary(req.file.buffer);
+
+        const createdLesson = await Lesson.create({
+            moduleId,
+            title,
+            overview,
+            duration,
+            order,
+            videoUrl: uploadResult.secure_url,
+            videoPublicId: uploadResult.public_id
+        })
+
+        return res
+        .status(200)
+        .json({message: "lesson created successfully", data: createdLesson})
+
+    } catch (error) {
+        return res
+        .status(500)
+        .json({message: "error creating lesson", error: error.message})
+    }
+}
+
+export const getModuleWithLessons = async (req, res) => {
+    try {
+        const {moduleId} = req.params
+        const module = await Module.findById(moduleId);
+        if (!module) {
+            return res
+            .status(400).json({message: "module not found"})
+        }
+        const moduleLessons = await Lesson.find({moduleId: moduleId}).sort({order: 1 })
+        return res
+        .status(200).json({message: "Lessons fetched successfully", data: {module, moduleLessons, totalLessons: moduleLessons.length}})
+    } catch (error) {
+        return res
+        .status(500).json({message: "error fetching Lessons", error: error.message})
     }
 }
 
@@ -70,6 +127,10 @@ export const getModuleById = async (req, res) => {
     try {
         const {id} = req.params;
         const requestedModule = await Module.findById(id);
+        if (!requestedModule) {
+            return res
+            .status(404).json({message: "module not found"})
+        }
 
         return res
         .status(200)
@@ -88,7 +149,12 @@ export const updateModule = async (req, res) => {
         const updatedModule = await Module.findByIdAndUpdate(id, req.body,{
             new: true,
             runValidators: true
-        })
+        });
+        if (!updatedModule) {
+            return res
+            .status(404).json({message: "module not found"})
+        }
+        
         return res
         .status(200)
         .json({message: "module updated successfully", data: updatedModule})
@@ -103,14 +169,24 @@ export const updateModule = async (req, res) => {
 export const deleteModule = async (req, res) => {
     try {
         const {id} = req.params
-        const deletedModule = await Module.findByIdAndDelete(id);
+        const module = await Module.findById(id);
+        if(!module) {
+            return res
+            .status(404).json({message: "module not found"});
+        }
+        const lessons = await Lesson.find({moduleId: id});
+        for (const lesson of lessons) {
+            await deleteFromCloudinary(lesson.videoPublicId);
+        }
+        const deletedLessons = await Lesson.deleteMany({moduleId: id});
+        const deletedModule = await Module.findByIdAndDelete(id)
         return res
         .status(200)
-        .json({message: "module deleted successfully", data: deletedModule})
+        .json({message: "module deleted successfully", data: {deletedLessons, deletedModule}})
     } catch (error) {
         return res
         .status(500)
-        .json({message: "error deleting user", error: error.message})
+        .json({message: "error deleting module", error: error.message})
     }
 }
